requirements.freeze.before_maad.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
attrs==26.1.0
audioread==3.1.0
av==17.0.0
braceexpand==0.1.7
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.3
charset-normalizer==3.4.7
click==8.2.1
contourpy==1.3.3
ctranslate2==4.7.1
cycler==0.12.1
decorator==5.2.1
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
ftfy==6.3.1
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.80.0
h11==0.16.0
h5py==3.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.2
idna==3.11
ImageIO==2.37.3
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.14.0
kiwisolver==1.5.0
laion_clap==1.1.7
lazy-loader==0.5
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.47.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mistune==3.2.0
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
namex==0.1.0
networkx==3.6.1
numba==0.65.0
numpy==1.26.4
onnxruntime==1.24.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
panns-inference==0.1.1
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
pooch==1.9.0
progressbar==2.5
protobuf==5.29.6
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
pyparsing==3.3.2
pypdf==6.10.2
pytest==9.0.3
python-dateutil==2.9.0.post0
PyWavelets==1.9.0
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
reportlab==4.3.1
requests==2.33.1
resampy==0.4.3
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scikit-maad==1.5.2
scipy==1.14.1
sentry-sdk==2.58.0
setuptools==80.10.2
shellingham==1.5.4
silero-vad==5.1.2
six==1.17.0
smmap==5.0.3
soundfile==0.12.1
soxr==1.0.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-hub==0.16.1
termcolor==3.3.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2026.4.11
tokenizers==0.22.2
torch==2.11.0
torchaudio==2.11.0
torchlibrosa==0.1.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.26.0
wcwidth==0.6.0
webdataset==1.0.2
Werkzeug==3.1.8
wget==3.2
wheel==0.46.3
wrapt==2.1.2
